HW7/requirements.txt
pillow
numpy
matplotlib
scipy